test: make presentationFormatTester more stable
In the AreaCircle/explanation/base/ test every so
often page 28 would fail in the next/prev testing.
Sometimes it would pass (usually when the number
of tests was reduced), but it would often fail.
Adding a small delay for after the navigation seems to make
it more stable. Don't know why.